classic essay, no ostis

+++ src/main/java/bsuir/ai/sostis/model/Sentence.java
@@ -18,5 +18,5 @@ public class Sentence {
     Integer number;
     String text;
     List<Word> words;
-    Integer rang;
+    Double rang;
 }
